backend/requirements.txt
flask
flask-cors
dwani
python-dotenv
